Fix GET issues
